Refactor namespaces and add new model classes
Reformatted `PackageReference` elements in `API.csproj` for better readability. Updated `using` directives in `PackagesController.cs`, `IPackageService`, `PackageService`, and `PackagesDomain` to reflect new namespaces.

Moved `Customer`, `LineOfCredit`, `Package`, `Catalog`, `Category`, and `Item` classes to appropriate namespaces. Added new `Cart`, `Role`, `User`, `Branch`, `City`, `Order`, `Region`, `Supplier`, and `Warehouse` classes under respective namespaces.

Minor formatting update in `PremiumCustomer` class.
