+++ API/Models/Customers/Cart.cs
@@ -1,0 +1,6 @@
+﻿namespace API.Models.Customers
+{
+    public class Cart
+    {
+    }
+}
