+++ API/Models/IntAdmin/Category.cs
@@ -1,8 +1,8 @@
-namespace API.Models;
+namespace API.Models.IntAdmin;
 
 public class Category
 {
     public string Name { get; set; }
-    
+
     public List<Item> Items { get; set; } = new List<Item>();
 }
