+++ API/Models/IntAdmin/Item.cs
@@ -1,10 +1,10 @@
-namespace API.Models;
+namespace API.Models.IntAdmin;
 
 public class Item
 {
     public string Sku { get; set; }
-    
+
     public decimal Price { get; set; }
-    
+
     public decimal Discount { get; set; }
 }
