+++ API/Models/IntAdmin/User.cs
@@ -1,0 +1,6 @@
+﻿namespace API.Models.IntAdmin
+{
+    public class User
+    {
+    }
+}
