+++ API/Models/Logistics/City.cs
@@ -1,0 +1,6 @@
+﻿namespace API.Models.Logistics
+{
+    public class City
+    {
+    }
+}
