+++ API/Models/Logistics/Region.cs
@@ -1,0 +1,6 @@
+﻿namespace API.Models.Logistics
+{
+    public class Region
+    {
+    }
+}
