+++ API/Models/Logistics/Supplier.cs
@@ -1,0 +1,6 @@
+﻿namespace API.Models.Logistics
+{
+    public class Supplier
+    {
+    }
+}
