+++ API/Services/IPackageService.cs
@@ -1,4 +1,6 @@
 using API.Models;
+using API.Models.Customers;
+using API.Models.IntAdmin;
 
 namespace API.Services
 {
